completed refactor of file structure
