+++ components/subreddits/dataDisplay.tsx
