+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.scss';
 import { AppProps } from 'next/app';
-import Layout from '../components/layout';
+import Layout from '../components/layout/layout';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
