+++ pages/users.tsx
@@ -1,0 +1,5 @@
+const Users = () => {
+  return <div>Users Page</div>;
+};
+
+export default Users;
